+++ Auction/Domain/UseCases/Extensions/DetailProductUsecase.swift
@@ -14,7 +14,7 @@ final class DetailProductUsecase{
     }
 }
 extension DetailProductUsecase:Pr_DetailProductUsecase{
-    func returnDetailProduct(productId: Int8) -> Observable<Result<DetailProduct, HTTPError>> {
+    func returnDetailProduct(productId: Int) -> Observable<Result<DetailProduct, HTTPError>> {
         detailProductRepository.httpDetailProduct(productId: productId)
     }
 }
