+++ Auction/Domain/UseCases/Protocol/Pr_CurrentProductPriceUsecase.swift
@@ -8,6 +8,6 @@
 import Foundation
 import RxSwift
 protocol Pr_CurrentProductPriceUsecase{
-    func returnCurrentProductPrice(productId:Int8) -> Observable<Result<CurrentProductPrice,HTTPError>>
+    func returnCurrentProductPrice(productId:Int) -> Observable<Result<CurrentProductPrice,HTTPError>>
     func returnStreamCurrentProductPrice() -> Observable<Result<[StreamPrice],StreamError>>
 }
