+++ Auction/Presentation/DetailProducts/DetailProductViewController/DetailProductPriceView/ViewModel/Pr_DetailProductPriceViewModel.swift
@@ -11,5 +11,5 @@ protocol Pr_DetailProductPriceViewModel{
     var updownObservable:Observable<ProductUpDown>{get}
     var priceObservable:Observable<String>{get}
     var beforePriceObservable:Observable<String>{get}
-    var requestDataObserver:AnyObserver<Int8>{get}
+    var requestDataObserver:AnyObserver<Int>{get}
 }
